editor/editor.py
```python
import os
from moviepy.editor import VideoFileClip, concatenate_videoclips
from PIL import Image

class VideoEditor:
    def __init__(self, video_path):
        self.video_path = video_path
        self.video_clip = VideoFileClip(video_path)

    def cut(self, start_time, end_time):
        return self.video_clip.subclip(start_time, end_time)

    def trim(self, start_time, end_time):
        trimmed_clip = self.video_clip.subclip(start_time, end_time)
        output_path = os.path.join(EXPORT_FOLDER, 'trimmed_' + os.path.basename(self.video_path))
        trimmed_clip.write_videofile(output_path, codec="libx264", audio_codec="aac")
        return output_path

    def crop(self, x1, y1, x2, y2):
        cropped_clip = self.video_clip.crop(x1=x1, y1=y1, x2=x2, y2=y2)
        output_path = os.path.join(EXPORT_FOLDER, 'cropped_' + os.path.basename(self.video_path))
        cropped_clip.write_videofile(output_path, codec="libx264", audio_codec="aac")
        return output_path

    def rotate(self, angle):
        rotated_clip = self.video_clip.rotate(angle)
        output_path = os.path.join(EXPORT_FOLDER, 'rotated_' + os.path.basename(self.video_path))
        rotated_clip.write_videofile(output_path, codec="libx264", audio_codec="aac")
        return output_path

    def add_text(self, text, position, fontsize=24, color='white', font='Arial'):
        from moviepy.editor import TextClip
        text_clip = TextClip(text, fontsize=fontsize, color=color, font=font)
        text_clip = text_clip.set_position(position).set_duration(self.video_clip.duration)
        video_with_text = concatenate_videoclips([self.video_clip, text_clip])
        output_path = os.path.join(EXPORT_FOLDER, 'text_added_' + os.path.basename(self.video_path))
        video_with_text.write_videofile(output_path, codec="libx264", audio_codec="aac")
        return output_path

    def add_effect(self, effect_name):
        # Placeholder for effect logic
        pass

    def adjust_brightness(self, level):
        # Placeholder for brightness adjustment logic
        pass

    def adjust_contrast(self, level):
        # Placeholder for contrast adjustment logic
        pass

    def generate_thumbnail(self):
        frame = self.video_clip.get_frame(1)
        image = Image.fromarray(frame)
        thumbnail_path = os.path.join(THUMBNAIL_FOLDER, 'thumbnail_' + os.path.basename(self.video_path).split('.')[0] + '.png')
        image.save(thumbnail_path)
        return thumbnail_path

    def close(self):
        self.video_clip.close()
```